Finsh the version 1

+++ supervisor/strategy.go
@@ -2,6 +2,6 @@ package supervisor
 
 // 要求在给定事件间隔期间，不能失败超过Number次
 type Strategy struct {
-	TimeInterval int
+	TimeInterval int64
 	Number       int
 }
